# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  projectType?: string | null;
  /** @minLength 1 */
  message: string;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  projectType?: string | null;
  message: string;
  createdAt: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminLoginResponse {
  token: string;
}

export interface ErrorResponse {
  error: string;
}