# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact form operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact form
      description: Accepts a contact inquiry from a website visitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Inquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      description: Validates admin password and returns a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/contacts:
    get:
      operationId: listContacts
      tags: [admin]
      summary: List all contact submissions
      description: Returns all contact form submissions, newest first
      parameters:
        - in: header
          name: x-admin-token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactSubmission"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ContactInput:
      type: object
      required:
        - name
        - email
        - message
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        company:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        projectType:
          type: ["string", "null"]
        message:
          type: string
          minLength: 1
    ContactSubmission:
      type: object
      required:
        - id
        - name
        - email
        - message
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        company:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        projectType:
          type: ["string", "null"]
        message:
          type: string
        createdAt:
          type: string
          format: date-time
    AdminLoginInput:
      type: object
      required:
        - password
      properties:
        password:
          type: string
    AdminLoginResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string